Reorganize Headers & Classes in tools folder
