+++ menrva/app/build.gradle
@@ -32,7 +32,6 @@ android {
     sourceSets {
         main {
             jni.srcDirs = ['src/main/cpp/jni/', 'src/androidTest/cpp/jni/']
-            jniLibs.srcDirs = ['../../libs/fftw/out/', '../../libs/kissfft/out/']
         }
     }
     compileOptions {
