+++ menrva/app/src/main/cpp/log/LoggerBase.cpp
