+++ menrva/app/src/main/cpp/log/LoggerBase.h
