+++ menrva/app/src/main/cpp/log/LoggingBase.h
